Add files via upload
RPG Experiments ; Learning to Code.

hit_func.py  is an attempt at a randomised hit rate (miss,parry,hit,criticalHit). Calculated with the sum of a users or npc stat and a weapon proficiency stat, gained through extended use.........(needs linking to sword.py for a damage output)


sword.py is an attempt at a weapon class and "mastery" system, with a weapon use counter that would go toward a proficiency score. 
I was hoping to setup a weapon condition stat the you would have to maintain. It would be linked to the weapon use counter. (needs to be upgradable for bigger weapon heath pool and damage output. 

comments welcome. 
wish me luck :D